pyproject.toml
[project]
name = "aerospike-py"
dynamic = ["version"]
description = "Aerospike Python Client - PyO3 + Rust binding"
readme = "README.md"
authors = [
    { name = "KimSoungRyoul", email = "KimSoungRyoul@gmail.com" }
]
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = []
keywords = ["aerospike", "database", "nosql", "client", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Rust",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
    "Typing :: Typed",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]

[project.urls]
Homepage = "https://github.com/KimSoungRyoul/aerospike-py"
Repository = "https://github.com/KimSoungRyoul/aerospike-py"
Issues = "https://github.com/KimSoungRyoul/aerospike-py/issues"

[project.optional-dependencies]
numpy = ["numpy>=2.0"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]
otel-export = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20", "opentelemetry-exporter-otlp>=1.20"]

[dependency-groups]
dev = ["pytest", "pytest-asyncio", "ruff", "maturin>=1.9,<2", "numpy>=2.0", "pyright>=1.1"]
bench = ["aerospike>=17.0.0", "numpy>=2.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["maturin>=1.9,<2"]
build-backend = "maturin"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # line length (handled by formatter)
    "UP007",   # use X | Y for union types (requires 3.10+ syntax everywhere)
    "UP045",   # Optional[X] → X | None (same as UP007, keep Optional for readability)
    "UP037",   # remove quotes from type annotations in stubs
    "UP035",   # import from collections.abc (keep typing imports for compatibility)
    "RUF022",  # unsorted __all__ (intentionally grouped by category)
]

[tool.ruff.lint.per-file-ignores]
"benchmark/*" = ["B023", "B905", "SIM105", "RUF002", "RUF059", "RUF100", "I001"]
"tests/*" = ["B011", "B007", "B905", "SIM105", "RUF059", "RUF100"]
"examples/sample-fastapi/**" = ["B008", "TC001", "TC002"]
"src/aerospike_py/__init__.py" = ["I001", "RUF100"]
"src/aerospike_py/exception.py" = ["I001"]
"*.pyi" = ["I001"]

[tool.ruff.lint.isort]
known-first-party = ["aerospike_py"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src/aerospike_py"]
reportMissingModuleSource = false
venvPath = "."
venv = ".venv"

[tool.maturin]
python-source = "src"
module-name = "aerospike_py._aerospike"
manifest-path = "rust/Cargo.toml"
features = ["pyo3/extension-module"]
include = ["LICENSE"]
